requirements.txt
streamlit
python-dotenv
google-generativeai
langchain-google-genai
faiss-cpu